requirements.txt
opencv_python==4.5.3.56
Pillow==9.2.0
python-dotenv==0.21.0
python_telegram_bot==20.0a2
pytz==2021.3
selenium==3.141.0